Add test_heap to test
